refactor(app): split sign-in/sign-up + delete useAuthFlow consolidation
Undo the 716-LOC useAuthFlow hook + mode-switching components. Each route
now owns its own state and calls Clerk Future API directly:
useSignIn().signIn.{emailCode,sso,finalize}, useSignUp().signUp.{create,
verifications,sso,finalize}. Metadata moves to thin segment layouts since
'use client' pages cannot export metadata. Shrinks _lib/search-params.ts
to error+ticket fields only. Net change: -1442 LOC.

Sentry breadcrumb strings preserved verbatim (Migration Notes).
Per plan/2026-05-14-auth-signin-signup-split.md Phase 3.
