--- apps/app/src/app/(auth)/_hooks/use-auth-flow.ts
@@ -1,716 +1,0 @@
-"use client";
-
-import { toast } from "@repo/ui/components/ui/sonner";
-import { useAuth, useClerk, useSignIn, useSignUp } from "@vendor/clerk/client";
-import type { OAuthStrategy } from "@vendor/clerk/types";
-import * as React from "react";
-import {
-  authErrorMessage,
-  mapOAuthClerkError,
-  mapOtpClerkError,
-} from "./auth-errors";
-import { authBreadcrumb, authSpan } from "./auth-telemetry";
-
-const SUCCESS_REDIRECT = "/account/welcome";
-
-type AuthMode = "sign-in" | "sign-up";
-type AuthStep = "email" | "code" | "activate";
-
-interface UseAuthFlowInput {
-  email?: string | null;
-  mode: AuthMode;
-  onWaitlistError?: () => void;
-  step: AuthStep;
-  ticket?: string | null;
-  token?: string | null;
-}
-
-interface OAuthSlice {
-  initiate: (strategy: OAuthStrategy) => Promise<void>;
-  loading: boolean;
-}
-
-interface OtpSlice {
-  code: string;
-  email: string | null;
-  error: string | null;
-  isInitializing: boolean;
-  isRedirecting: boolean;
-  isResending: boolean;
-  isVerifying: boolean;
-  onCodeChange: (v: string) => void;
-  onResend: () => Promise<void>;
-  onReset: () => void;
-}
-
-interface ActivateSlice {
-  error: string | null;
-}
-
-interface UseAuthFlowReturn {
-  activate: ActivateSlice;
-  oauth: OAuthSlice;
-  otp: OtpSlice;
-}
-
-// Clerk passes { session, decorateUrl } to the navigate callback in
-// finalize({ navigate }). decorateUrl appends ITP cookie-refresh params for
-// Safari third-party cookie environments. We hard-nav (window.location.href)
-// because finalize calls navigate AFTER the session is active and we want a
-// full reload to pick up any server-rendered state on /account/welcome.
-interface NavigateArgs {
-  decorateUrl: (url: string) => string;
-}
-
-export function useAuthFlow(input: UseAuthFlowInput): UseAuthFlowReturn {
-  const { mode, step, email, ticket, token, onWaitlistError } = input;
-  const { signIn } = useSignIn();
-  const { signUp } = useSignUp();
-  const clerk = useClerk();
-  // useAuth gives a reactive isLoaded that flips when Clerk has hydrated the
-  // browser client (clerk.client.signIn becomes safe to call). useSignIn's own
-  // isLoaded was returning undefined in dev with the Future-API toggle on.
-  const { isLoaded: isClerkLoaded } = useAuth();
-
-  const navigateToSuccess = React.useCallback(
-    async ({ decorateUrl }: NavigateArgs) => {
-      window.location.href = decorateUrl(SUCCESS_REDIRECT);
-    },
-    []
-  );
-
-  const handleWaitlist = React.useCallback(() => {
-    if (onWaitlistError) {
-      onWaitlistError();
-      return;
-    }
-    const target = mode === "sign-in" ? "/sign-in" : "/sign-up";
-    // .replace() not .href: the user came from /sign-{in,up}?step=code (OTP)
-    // or from /sign-{in,up}/sso-callback. Both terminal — back-button should
-    // skip them, not re-enter (which would re-fire create() / handleRedirect-
-    // Callback() and bounce forward again).
-    window.location.replace(`${target}?errorCode=waitlist`);
-  }, [onWaitlistError, mode]);
-
-  // --- OAuth slice ---
-  const [oauthLoading, setOauthLoading] = React.useState(false);
-
-  // When the user clicks an OAuth button, we flip oauthLoading→true and the
-  // Clerk SDK navigates the tab to the IdP. If the user then hits Back without
-  // picking an account, modern Chrome restores /sign-in from the bfcache with
-  // the React tree intact — oauthLoading stays true and every OAuth button on
-  // the page is left disabled, so subsequent clicks land on a disabled button
-  // and do nothing.
-  //
-  // Defense in depth:
-  // 1. pagehide: reset oauthLoading→false BEFORE the page is bfcached, so the
-  //    snapshot itself has the correct state. The user never sees the
-  //    transition (page is unloading).
-  // 2. pageshow: also reset on every page-show event (initial load, bfcache
-  //    restore, plain back-nav). We drop the persisted=true guard because
-  //    setting to false when already false is a no-op, and some browser/SDK
-  //    combinations don't deliver persisted=true reliably for partial OAuth
-  //    flows. Cheaper to just always reset.
-  React.useEffect(() => {
-    const reset = () => setOauthLoading(false);
-    window.addEventListener("pagehide", reset);
-    window.addEventListener("pageshow", reset);
-    return () => {
-      window.removeEventListener("pagehide", reset);
-      window.removeEventListener("pageshow", reset);
-    };
-  }, []);
-
-  const handleOAuthMapped = React.useCallback(
-    (
-      mapped: ReturnType<typeof mapOAuthClerkError>,
-      strategy: OAuthStrategy,
-      label: string
-    ) => {
-      if (mapped.kind === "code" && mapped.errorCode === "waitlist") {
-        authBreadcrumb(`OAuth blocked by waitlist (${label})`, "warning", {
-          strategy,
-        });
-        handleWaitlist();
-        return;
-      }
-      if (mapped.kind === "redirect") {
-        window.location.href = mapped.target;
-        return;
-      }
-      if (mapped.kind === "inline") {
-        toast.error(mapped.message);
-      }
-    },
-    [handleWaitlist]
-  );
-
-  const initiateOAuth = React.useCallback(
-    async (strategy: OAuthStrategy): Promise<void> => {
-      setOauthLoading(true);
-      // Breadcrumb message preserved verbatim from oauth-button.tsx so
-      // existing Sentry dashboards/alerts keyed on the string still match.
-      authBreadcrumb("OAuth sign-in initiated", "info", { strategy, mode });
-
-      try {
-        if (mode === "sign-up" && ticket) {
-          // Ticket-OAuth flow: signUp.create({ ticket }) creates the resource;
-          // a follow-up call PATCHes it with the OAuth strategy + legal_accepted
-          // and gets the IdP redirect URL.
-          //
-          // The Future API's signUp.sso() is broken in clerk-js@6.8.0 when
-          // called after signUp.create({ticket}): it POSTs to the *collection*
-          // URL /v1/client/sign_ups (with ?_method=PATCH) instead of the
-          // resource URL /v1/client/sign_ups/{id}, returning 405. Drop to the
-          // legacy clerk.client.signUp.authenticateWithRedirect with
-          // continueSignUp:true, which routes through SignUp.update() and
-          // PATCHes /v1/client/sign_ups/{id} correctly. Mirrors the activate
-          // slice workaround at use-auth-flow.ts:553.
-          const { error: createError } = await authSpan(
-            "auth.ticket.create",
-            { mode, strategy },
-            () => signUp.create({ ticket })
-          );
-          if (createError) {
-            handleOAuthMapped(
-              mapOAuthClerkError(createError),
-              strategy,
-              "ticket create"
-            );
-            setOauthLoading(false);
-            return;
-          }
-
-          try {
-            await authSpan("auth.oauth.initiate", { mode, strategy }, () =>
-              clerk.client.signUp.authenticateWithRedirect({
-                strategy,
-                redirectUrl: `/sign-up/sso-callback?__clerk_ticket=${encodeURIComponent(ticket)}`,
-                redirectUrlComplete: SUCCESS_REDIRECT,
-                continueSignUp: true,
-                legalAccepted: true,
-              })
-            );
-          } catch (ssoError) {
-            handleOAuthMapped(
-              mapOAuthClerkError(ssoError),
-              strategy,
-              "ticket flow"
-            );
-            setOauthLoading(false);
-          }
-          return;
-        }
-
-        const callbackUrl =
-          mode === "sign-in"
-            ? "/sign-in/sso-callback"
-            : "/sign-up/sso-callback";
-
-        // Future API's signIn.sso() / signUp.sso() silently no-op against a
-        // resource that already has a terminal verification state (e.g. a
-        // sticky sign_up_restricted_waitlist on signIn.firstFactorVerification
-        // after a prior OAuth round). Zero network traffic, the spinner just
-        // hangs. The Future API exposes no `continueSign{In,Up}` escape hatch
-        // and reset() doesn't clear the cookie-encoded state — only the legacy
-        // authenticateWithRedirect with continueSign{In,Up}:false forces
-        // clerk-js to POST a fresh /v1/client/sign_{ins,ups} resource rather
-        // than PATCH the stale one. Mirrors the existing Bug D workaround for
-        // the ticket flow above (legacy authenticateWithRedirect there too).
-        //
-        // Standard sign-up SSO intentionally omits legalAccepted —
-        // sign-up/sso-callback/page.tsx runs a patch effect that applies it
-        // post-callback. Adding legalAccepted here would race that patch.
-        try {
-          if (mode === "sign-in") {
-            clerk.client?.resetSignIn?.();
-            await authSpan("auth.oauth.initiate", { mode, strategy }, () =>
-              clerk.client.signIn.authenticateWithRedirect({
-                strategy,
-                redirectUrl: callbackUrl,
-                redirectUrlComplete: SUCCESS_REDIRECT,
-                continueSignIn: false,
-              })
-            );
-          } else {
-            clerk.client?.resetSignUp?.();
-            await authSpan("auth.oauth.initiate", { mode, strategy }, () =>
-              clerk.client.signUp.authenticateWithRedirect({
-                strategy,
-                redirectUrl: callbackUrl,
-                redirectUrlComplete: SUCCESS_REDIRECT,
-                continueSignUp: false,
-              })
-            );
-          }
-        } catch (ssoError) {
-          handleOAuthMapped(mapOAuthClerkError(ssoError), strategy, "sso");
-          setOauthLoading(false);
-        }
-      } catch {
-        toast.error("An unexpected error occurred");
-        setOauthLoading(false);
-      }
-    },
-    [mode, ticket, clerk, signUp, handleOAuthMapped]
-  );
-
-  // --- OTP slice ---
-  const [code, setCode] = React.useState("");
-  const [otpError, setOtpError] = React.useState<string | null>(null);
-  const [isVerifying, setIsVerifying] = React.useState(false);
-  const [isRedirecting, setIsRedirecting] = React.useState(false);
-  const [isResending, setIsResending] = React.useState(false);
-  const [isInitializing, setIsInitializing] = React.useState(step === "code");
-  const [resolvedEmail, setResolvedEmail] = React.useState<string | null>(
-    email ?? null
-  );
-
-  const hasInitRef = React.useRef(false);
-  const verifyingCodeRef = React.useRef<string | null>(null);
-
-  const handleOtpClerkError = React.useCallback(
-    (err: unknown): { success: boolean } => {
-      if (!err) {
-        return { success: false };
-      }
-      const mapped = mapOtpClerkError(err);
-      if (mapped.kind === "success") {
-        return { success: true };
-      }
-      if (mapped.kind === "redirect") {
-        // session_exists → /account/welcome. .replace so back-button doesn't
-        // re-enter the verify step.
-        window.location.replace(mapped.target);
-        return { success: false };
-      }
-      if (mapped.kind === "code") {
-        if (mapped.errorCode === "waitlist") {
-          handleWaitlist();
-          return { success: false };
-        }
-        setOtpError(authErrorMessage(mapped.errorCode));
-        return { success: false };
-      }
-      setOtpError(mapped.message);
-      return { success: false };
-    },
-    [handleWaitlist]
-  );
-
-  // Init effect: send the OTP (or apply the invitation ticket) exactly once
-  // when step === "code".
-  //
-  // EmailForm now calls create + sendCode on submit (so the user never sees
-  // OTP UI flash before bouncing to an error banner). On the happy path the
-  // Clerk in-flight resource is already primed for this email by the time
-  // we get here — we detect that and skip the send so a second OTP email
-  // doesn't go out / trip too_many_requests. The cold path (direct nav to
-  // ?step=code, cookie-cleared refresh, etc.) falls through to the regular
-  // send.
-  React.useEffect(() => {
-    if (step !== "code" || hasInitRef.current) {
-      return;
-    }
-    hasInitRef.current = true;
-
-    if (mode === "sign-in" && !email) {
-      setOtpError("Missing email. Please start over.");
-      setIsInitializing(false);
-      return;
-    }
-
-    const sameEmail = (
-      a: string | null | undefined,
-      b: string | null | undefined
-    ) => !!a && !!b && a.toLowerCase() === b.toLowerCase();
-
-    const signInPrimed =
-      mode === "sign-in" &&
-      signIn.status === "needs_first_factor" &&
-      sameEmail(signIn.identifier, email) &&
-      signIn.firstFactorVerification?.status === "unverified" &&
-      signIn.firstFactorVerification?.strategy === "email_code";
-
-    const signUpPrimed =
-      mode === "sign-up" &&
-      signUp.status === "missing_requirements" &&
-      sameEmail(signUp.emailAddress, email) &&
-      signUp.verifications?.emailAddress?.status === "unverified" &&
-      signUp.verifications?.emailAddress?.strategy === "email_code";
-
-    if (signInPrimed || signUpPrimed) {
-      authBreadcrumb("OTP init skipped — resource pre-primed", "info", {
-        mode,
-      });
-      setResolvedEmail(
-        mode === "sign-in"
-          ? (signIn.identifier ?? email ?? null)
-          : (signUp.emailAddress ?? email ?? null)
-      );
-      setIsInitializing(false);
-      return;
-    }
-
-    async function init() {
-      if (mode === "sign-up" && ticket) {
-        // Clerk's invitation flow does NOT auto-populate signUp.emailAddress
-        // from the ticket — without explicit emailAddress, the SignUp resource
-        // has missingFields=["email_address"] and verifications.sendEmailCode()
-        // fails with "Email address missing on Sign Up Preparation".
-        // strategy: "ticket" disambiguates from the no-ticket sign-up path.
-        const { error: createError } = await signUp.create({
-          strategy: "ticket",
-          ticket,
-          emailAddress: email ?? undefined,
-          legalAccepted: true,
-        });
-        if (createError) {
-          const mapped = mapOtpClerkError(createError);
-          // Ticket-consume rejections (revoked, already-redeemed, expired,
-          // generic) happen AFTER the URL has transitioned to ?step=code, so
-          // surfacing them inline would layer the error over the "We sent a
-          // verification code" UI — confusing copy. Reset to the email step
-          // with ErrorBanner instead. The ticket is no longer usable; drop it.
-          // .replace on all error paths: the user is on /sign-up?step=code
-          // (terminal in this branch — ticket already consumed). Back-button
-          // should bypass it, not re-fire signUp.create.
-          if (mapped.kind === "redirect") {
-            window.location.replace(mapped.target);
-            return;
-          }
-          if (mapped.kind === "code") {
-            if (mapped.errorCode === "waitlist") {
-              handleWaitlist();
-              return;
-            }
-            window.location.replace(`/sign-up?errorCode=${mapped.errorCode}`);
-            return;
-          }
-          if (mapped.kind === "inline") {
-            const params = new URLSearchParams({ error: mapped.message });
-            window.location.replace(`/sign-up?${params.toString()}`);
-            return;
-          }
-          // mapped.kind === "success" — fall through to the status check.
-        }
-        if (signUp.status === "complete") {
-          setIsRedirecting(true);
-          await signUp.finalize({ navigate: navigateToSuccess });
-          return;
-        }
-        setResolvedEmail(signUp.emailAddress ?? email ?? null);
-
-        const { error: sendError } = await authSpan(
-          "auth.otp.send",
-          { mode },
-          () => signUp.verifications.sendEmailCode()
-        );
-        if (sendError) {
-          authBreadcrumb("OTP send failed", "error", {
-            code: sendError.code,
-            mode,
-          });
-          handleOtpClerkError(sendError);
-        } else {
-          authBreadcrumb("OTP code sent", "info", { mode, email });
-        }
-        return;
-      }
-
-      if (mode === "sign-in") {
-        const { error: sendError } = await authSpan(
-          "auth.otp.send",
-          { mode },
-          () => signIn.emailCode.sendCode({ emailAddress: email ?? undefined })
-        );
-        if (sendError) {
-          authBreadcrumb("OTP send failed", "error", {
-            code: sendError.code,
-            mode,
-          });
-          handleOtpClerkError(sendError);
-        } else {
-          authBreadcrumb("OTP code sent", "info", { mode, email });
-        }
-        return;
-      }
-
-      // mode === "sign-up" without ticket
-      const { error: createError } = await signUp.create({
-        emailAddress: email ?? undefined,
-        legalAccepted: true,
-      });
-      if (createError) {
-        handleOtpClerkError(createError);
-        return;
-      }
-      const { error: sendError } = await authSpan(
-        "auth.otp.send",
-        { mode },
-        () => signUp.verifications.sendEmailCode()
-      );
-      if (sendError) {
-        authBreadcrumb("OTP send failed", "error", {
-          code: sendError.code,
-          mode,
-        });
-        handleOtpClerkError(sendError);
-      } else {
-        authBreadcrumb("OTP code sent", "info", { mode, email });
-      }
-    }
-
-    init()
-      .catch(() =>
-        setOtpError("An unexpected error occurred. Please try again.")
-      )
-      .finally(() => setIsInitializing(false));
-  }, [
-    step,
-    mode,
-    ticket,
-    email,
-    signIn,
-    signUp,
-    handleOtpClerkError,
-    handleWaitlist,
-    navigateToSuccess,
-  ]);
-
-  // Verify the OTP code. Triggered from onCodeChange when the input reaches
-  // length 6 — no effect needed. verifyingCodeRef deduplicates concurrent
-  // submissions of the same code (e.g. re-render mid-flight).
-  const verifyOtp = React.useCallback(
-    async (codeValue: string) => {
-      authBreadcrumb("OTP verification attempt", "info", { mode });
-      setIsVerifying(true);
-      try {
-        if (mode === "sign-in") {
-          const { error: verifyError } = await authSpan(
-            "auth.otp.verify",
-            { mode },
-            () => signIn.emailCode.verifyCode({ code: codeValue })
-          );
-          if (verifyError) {
-            authBreadcrumb("OTP verification failed", "warning", {
-              code: verifyError.code,
-              mode,
-            });
-            const { success } = handleOtpClerkError(verifyError);
-            if (!success) {
-              verifyingCodeRef.current = null;
-              setIsVerifying(false);
-              return;
-            }
-          }
-          if (signIn.status === "complete") {
-            authBreadcrumb("OTP verified", "info", { mode });
-            setIsRedirecting(true);
-            await signIn.finalize({ navigate: navigateToSuccess });
-          } else {
-            verifyingCodeRef.current = null;
-            setOtpError(
-              "Sign-in could not be completed. Please try again or contact support."
-            );
-            setIsVerifying(false);
-          }
-          return;
-        }
-
-        const { error: verifyError } = await authSpan(
-          "auth.otp.verify",
-          { mode },
-          () => signUp.verifications.verifyEmailCode({ code: codeValue })
-        );
-        if (verifyError) {
-          authBreadcrumb("OTP verification failed", "warning", {
-            code: verifyError.code,
-            mode,
-          });
-          const { success } = handleOtpClerkError(verifyError);
-          if (!success) {
-            verifyingCodeRef.current = null;
-            setIsVerifying(false);
-            return;
-          }
-        }
-        if (signUp.status === "complete") {
-          authBreadcrumb("OTP verified", "info", { mode });
-          setIsRedirecting(true);
-          await signUp.finalize({ navigate: navigateToSuccess });
-        } else {
-          verifyingCodeRef.current = null;
-          setOtpError(
-            "Sign-up could not be completed. Please try again or contact support."
-          );
-          setIsVerifying(false);
-        }
-      } catch {
-        verifyingCodeRef.current = null;
-        setOtpError("An unexpected error occurred. Please try again.");
-        setIsVerifying(false);
-      }
-    },
-    [mode, signIn, signUp, handleOtpClerkError, navigateToSuccess]
-  );
-
-  const onCodeChange = React.useCallback(
-    (value: string) => {
-      setOtpError(null);
-      if (value.length < 6) {
-        verifyingCodeRef.current = null;
-      }
-      setCode(value);
-
-      // Auto-fire verify when the user reaches a full 6-digit code. Gated on
-      // !isInitializing so a paste during the initial sendCode() doesn't try
-      // to verify against a not-yet-ready resource. verifyingCodeRef prevents
-      // the same code from being submitted twice (e.g. React batching).
-      if (value.length !== 6 || isInitializing) {
-        return;
-      }
-      if (verifyingCodeRef.current === value) {
-        return;
-      }
-      verifyingCodeRef.current = value;
-      void verifyOtp(value);
-    },
-    [isInitializing, verifyOtp]
-  );
-
-  const onResend = React.useCallback(async () => {
-    // Don't race the init effect's initial sendCode — Clerk will return
-    // too_many_requests for the second concurrent send.
-    if (isInitializing) {
-      return;
-    }
-    setIsResending(true);
-    setOtpError(null);
-    authBreadcrumb("OTP resend requested", "info", { mode });
-    try {
-      const { error: sendError } = await authSpan(
-        "auth.otp.send",
-        { mode },
-        () =>
-          mode === "sign-in"
-            ? signIn.emailCode.sendCode({ emailAddress: email ?? undefined })
-            : signUp.verifications.sendEmailCode()
-      );
-      if (sendError) {
-        authBreadcrumb("OTP resend failed", "error", {
-          code: sendError.code,
-          mode,
-        });
-        handleOtpClerkError(sendError);
-      } else {
-        authBreadcrumb("OTP code resent", "info", { mode });
-        toast.success("Verification code sent to your email");
-        setCode("");
-        verifyingCodeRef.current = null;
-      }
-    } catch {
-      setOtpError("An unexpected error occurred. Please try again.");
-    }
-    setIsResending(false);
-  }, [isInitializing, mode, email, signIn, signUp, handleOtpClerkError]);
-
-  const onReset = React.useCallback(() => {
-    // .replace because user pressed "Back" — they want to start over, not
-    // build a history loop.
-    if (mode === "sign-in") {
-      window.location.replace("/sign-in");
-      return;
-    }
-    const ticketParam = ticket
-      ? `?__clerk_ticket=${encodeURIComponent(ticket)}`
-      : "";
-    window.location.replace(`/sign-up${ticketParam}`);
-  }, [mode, ticket]);
-
-  // --- Activate slice ---
-  const [activateError, setActivateError] = React.useState<string | null>(null);
-  const hasActivatedRef = React.useRef(false);
-
-  React.useEffect(() => {
-    if (step !== "activate" || !token || hasActivatedRef.current) {
-      return;
-    }
-    // Wait for Clerk to finish hydrating; clerk.client is undefined until then
-    // and clerk.client.signIn.create would throw "Cannot read 'signIn' of undefined".
-    if (!(isClerkLoaded && clerk?.client?.signIn)) {
-      return;
-    }
-    // Strict-mode + invitation-ticket guard: the underlying Clerk client
-    // consumes the ticket on first call. Without this ref, React 18 strict
-    // mode's double effect-fire in dev calls create() twice — the second call
-    // sees an already-consumed ticket and trips a spurious "Sign-in failed".
-    hasActivatedRef.current = true;
-
-    const ticketToken = token;
-    async function activate() {
-      authBreadcrumb("Session activation via ticket", "info", {});
-      // Clerk's Future API signIn proxy (useSignIn) does not propagate ticket-
-      // based sign-in state: signIn.ticket() and signIn.create({ strategy:
-      // "ticket" }) both leave the React signIn.status null even though the
-      // underlying Client creates a pending session. We drop to clerk.client
-      // for the call (which returns a usable status + createdSessionId) and
-      // promote via clerk.setActive.
-      try {
-        const result = await authSpan("auth.session.activate", { mode }, () =>
-          clerk.client.signIn.create({
-            strategy: "ticket",
-            ticket: ticketToken,
-          })
-        );
-        if (result.status === "complete" && result.createdSessionId) {
-          authBreadcrumb("Session activated", "info", {});
-          await clerk.setActive({ session: result.createdSessionId });
-          window.location.href = SUCCESS_REDIRECT;
-          return;
-        }
-        setActivateError("Sign-in failed. Please try again.");
-      } catch (err) {
-        // verification_already_verified can fire here on retry/refresh — if a
-        // session is already active, treat that as success.
-        const mapped = mapOtpClerkError(err);
-        if (mapped.kind === "success" && clerk.session) {
-          window.location.href = SUCCESS_REDIRECT;
-          return;
-        }
-        // Surface specific error copy when available (e.g. ticket_expired →
-        // "This invitation link has expired"). Falls back to the generic copy
-        // for the unmapped default branch.
-        if (mapped.kind === "inline" && mapped.message) {
-          setActivateError(mapped.message);
-          return;
-        }
-        if (mapped.kind === "code") {
-          setActivateError(authErrorMessage(mapped.errorCode));
-          return;
-        }
-        setActivateError("Sign-in failed. Please try again.");
-      }
-    }
-    activate().catch(() =>
-      setActivateError("Sign-in failed. Please try again.")
-    );
-  }, [step, token, mode, clerk, isClerkLoaded]);
-
-  return {
-    oauth: { initiate: initiateOAuth, loading: oauthLoading },
-    otp: {
-      code,
-      email: resolvedEmail,
-      error: otpError,
-      isInitializing,
-      isVerifying,
-      isRedirecting,
-      isResending,
-      onCodeChange,
-      onResend,
-      onReset,
-    },
-    activate: { error: activateError },
-  };
-}
