+++ apps/app/src/app/(auth)/sign-in/layout.tsx
@@ -1,0 +1,35 @@
+import { createMetadata } from "@vendor/seo/metadata";
+import type { Metadata } from "next";
+import type React from "react";
+
+export const metadata: Metadata = createMetadata({
+  title: "Sign In - Lightfast Auth",
+  description:
+    "Sign in to your Lightfast account to access the AI agent platform. Secure authentication portal for developers.",
+  openGraph: {
+    title: "Sign In - Lightfast Auth",
+    description:
+      "Sign in to your Lightfast account to access the AI agent platform.",
+    url: "https://lightfast.ai/sign-in",
+  },
+  twitter: {
+    title: "Sign In - Lightfast Auth",
+    description:
+      "Sign in to your Lightfast account to access the AI agent platform.",
+  },
+  alternates: {
+    canonical: "https://lightfast.ai/sign-in",
+  },
+  robots: {
+    index: true,
+    follow: false,
+  },
+});
+
+export default function SignInLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return children;
+}
