+++ apps/app/src/app/(auth)/sign-up/page.tsx
@@ -1,203 +1,503 @@
+"use client";
+
+import { Icons } from "@repo/ui/components/icons";
 import { Button } from "@repo/ui/components/ui/button";
-import { createMetadata } from "@vendor/seo/metadata";
+import { Checkbox } from "@repo/ui/components/ui/checkbox";
+import { Input } from "@repo/ui/components/ui/input";
+import { toast } from "@repo/ui/components/ui/sonner";
+import { useSignUp } from "@vendor/clerk/client";
+import type { OAuthStrategy } from "@vendor/clerk/types";
 import { Link as MicrofrontendLink } from "@vercel/microfrontends/next/client";
-import type { Metadata } from "next";
-import dynamic from "next/dynamic";
 import NextLink from "next/link";
-import { redirect } from "next/navigation";
-import type { SearchParams } from "nuqs/server";
+import { useSearchParams } from "next/navigation";
+import * as React from "react";
 import { env } from "~/env";
-import { EmailForm } from "../_components/email-form";
 import { ErrorBanner } from "../_components/error-banner";
-import { OAuthButton } from "../_components/oauth-button";
 import { SeparatorWithText } from "../_components/separator-with-text";
-import { loadSignUpSearchParams } from "../_lib/search-params";
+import { CodeVerificationUI } from "../_components/shared/code-verification-ui";
+import {
+  authErrorMessage,
+  mapOAuthClerkError,
+  mapOtpClerkError,
+} from "../_hooks/auth-errors";
+import { authBreadcrumb, authSpan } from "../_hooks/auth-telemetry";
+import { type AuthErrorCode, authErrorCodes } from "../_lib/search-params";
 
-const OTPIsland = dynamic(() =>
-  import("../_components/otp-island").then((m) => m.OTPIsland)
-);
+const SUCCESS_REDIRECT = "/account/welcome";
 
-function decodeTicketExpiry(ticket: string): Date | null {
-  try {
-    const segment = ticket.split(".")[1];
-    if (!segment) {
-      return null;
-    }
-    const payload = JSON.parse(
-      atob(segment.replace(/-/g, "+").replace(/_/g, "/"))
-    ) as { exp?: unknown };
-    return typeof payload.exp === "number"
-      ? new Date(payload.exp * 1000)
-      : null;
-  } catch {
+type View = "email" | "code";
+
+function parseErrorCode(value: string | null): AuthErrorCode | null {
+  if (!value) {
     return null;
   }
-}
-
-export const metadata: Metadata = createMetadata({
-  title: "Sign Up - Lightfast Auth",
-  description:
-    "Create your Lightfast account to access the AI agent platform. Secure sign-up portal for developers.",
-  openGraph: {
-    title: "Sign Up - Lightfast Auth",
-    description:
-      "Create your Lightfast account to access the AI agent platform.",
-    url: "https://lightfast.ai/sign-up",
-  },
-  twitter: {
-    title: "Sign Up - Lightfast Auth",
-    description:
-      "Create your Lightfast account to access the AI agent platform.",
-  },
-  alternates: {
-    canonical: "https://lightfast.ai/sign-up",
-  },
-  robots: {
-    index: true,
-    follow: false,
-  },
-});
-
-interface PageProps {
-  searchParams: Promise<SearchParams>;
-}
-
-export default async function SignUpPage({ searchParams }: PageProps) {
-  const { step, email, error, ticket, __clerk_ticket, errorCode } =
-    await loadSignUpSearchParams(searchParams);
-
-  // Support both ?ticket= (nuqs) and ?__clerk_ticket= (Clerk invitation URL)
-  const invitationTicket = ticket ?? __clerk_ticket ?? null;
-
-  // Required-param guard: a `?step=code` URL without `email` renders an empty
-  // page (`<OTPIsland>` is gated on email). Reset to the email step.
-  if (step === "code" && !email) {
-    redirect(
-      invitationTicket
-        ? `/sign-up?__clerk_ticket=${encodeURIComponent(invitationTicket)}`
-        : "/sign-up"
-    );
-  }
-
-  const signUpBaseUrl = invitationTicket
-    ? `/sign-up?__clerk_ticket=${encodeURIComponent(invitationTicket)}`
-    : "/sign-up";
-
-  const invitationExpiry = invitationTicket
-    ? decodeTicketExpiry(invitationTicket)
+  return (authErrorCodes as readonly string[]).includes(value)
+    ? (value as AuthErrorCode)
     : null;
+}
 
-  const hasError = !!(error ?? errorCode);
+export default function SignUpPage() {
+  const { signUp } = useSignUp();
+  const searchParams = useSearchParams();
+  const errorParam = searchParams.get("error");
+  const errorCode = parseErrorCode(searchParams.get("errorCode"));
+  const hasError = !!(errorParam ?? errorCode);
+
+  const [view, setView] = React.useState<View>("email");
+  const [email, setEmail] = React.useState("");
+  const [legalAccepted, setLegalAccepted] = React.useState(false);
+  const [legalError, setLegalError] = React.useState<string | null>(null);
+  const [submittedEmail, setSubmittedEmail] = React.useState<string | null>(
+    null
+  );
+  const [submitting, setSubmitting] = React.useState(false);
+  const [code, setCode] = React.useState("");
+  const [otpError, setOtpError] = React.useState<string | null>(null);
+  const [isVerifying, setIsVerifying] = React.useState(false);
+  const [isRedirecting, setIsRedirecting] = React.useState(false);
+  const [isResending, setIsResending] = React.useState(false);
+  const [oauthLoading, setOauthLoading] = React.useState(false);
+
+  const verifyingCodeRef = React.useRef<string | null>(null);
+
+  React.useEffect(() => {
+    const reset = () => setOauthLoading(false);
+    window.addEventListener("pagehide", reset);
+    window.addEventListener("pageshow", reset);
+    return () => {
+      window.removeEventListener("pagehide", reset);
+      window.removeEventListener("pageshow", reset);
+    };
+  }, []);
+
+  const handleWaitlist = React.useCallback(() => {
+    window.location.replace("/sign-up?errorCode=waitlist");
+  }, []);
+
+  const errorPathFor = React.useCallback(
+    (params: { errorCode?: string; error?: string }) => {
+      const search = new URLSearchParams();
+      if (params.errorCode) {
+        search.set("errorCode", params.errorCode);
+      } else if (params.error) {
+        search.set("error", params.error);
+      }
+      return `/sign-up?${search.toString()}`;
+    },
+    []
+  );
+
+  const handleOtpClerkError = React.useCallback(
+    (err: unknown): { success: boolean } => {
+      if (!err) {
+        return { success: false };
+      }
+      const mapped = mapOtpClerkError(err);
+      if (mapped.kind === "success") {
+        return { success: true };
+      }
+      if (mapped.kind === "redirect") {
+        window.location.replace(mapped.target);
+        return { success: false };
+      }
+      if (mapped.kind === "code") {
+        if (mapped.errorCode === "waitlist") {
+          handleWaitlist();
+          return { success: false };
+        }
+        setOtpError(authErrorMessage(mapped.errorCode));
+        return { success: false };
+      }
+      setOtpError(mapped.message);
+      return { success: false };
+    },
+    [handleWaitlist]
+  );
+
+  const handleSubmitEmailError = React.useCallback(
+    (err: unknown) => {
+      const mapped = mapOtpClerkError(err);
+      if (mapped.kind === "redirect") {
+        window.location.href = mapped.target;
+        return;
+      }
+      if (mapped.kind === "code") {
+        window.location.replace(errorPathFor({ errorCode: mapped.errorCode }));
+        return;
+      }
+      if (mapped.kind === "inline") {
+        window.location.replace(errorPathFor({ error: mapped.message }));
+        return;
+      }
+      window.location.replace(errorPathFor({ error: "Authentication failed" }));
+    },
+    [errorPathFor]
+  );
+
+  const onSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
+    event.preventDefault();
+    const trimmed = email.trim();
+    if (!trimmed || submitting) {
+      return;
+    }
+    if (!legalAccepted) {
+      setLegalError(
+        "You must accept the Terms of Service and Privacy Policy to continue."
+      );
+      return;
+    }
+    setLegalError(null);
+    setSubmitting(true);
+
+    try {
+      const { error: createError } = await authSpan(
+        "auth.signup.create",
+        { mode: "sign-up" },
+        () =>
+          signUp.create({
+            emailAddress: trimmed,
+            legalAccepted: true,
+          })
+      );
+      if (createError) {
+        authBreadcrumb("Email submit rejected", "warning", {
+          mode: "sign-up",
+          code: createError.code,
+        });
+        handleSubmitEmailError(createError);
+        return;
+      }
+
+      if (signUp.status === "complete") {
+        await signUp.finalize({
+          navigate: ({ decorateUrl }) => {
+            window.location.href = decorateUrl(SUCCESS_REDIRECT);
+          },
+        });
+        return;
+      }
+
+      const { error: sendError } = await authSpan(
+        "auth.otp.send",
+        { mode: "sign-up" },
+        () => signUp.verifications.sendEmailCode()
+      );
+      if (sendError) {
+        authBreadcrumb("OTP send rejected", "warning", {
+          mode: "sign-up",
+          code: sendError.code,
+        });
+        handleSubmitEmailError(sendError);
+        return;
+      }
+
+      authBreadcrumb("OTP code sent (from EmailForm)", "info", {
+        mode: "sign-up",
+      });
+      setSubmittedEmail(trimmed);
+      setView("code");
+      setSubmitting(false);
+    } catch (err) {
+      handleSubmitEmailError(err);
+    }
+  };
+
+  const verifyOtp = React.useCallback(
+    async (codeValue: string) => {
+      authBreadcrumb("OTP verification attempt", "info", { mode: "sign-up" });
+      setIsVerifying(true);
+      try {
+        const { error: verifyError } = await authSpan(
+          "auth.otp.verify",
+          { mode: "sign-up" },
+          () => signUp.verifications.verifyEmailCode({ code: codeValue })
+        );
+        if (verifyError) {
+          authBreadcrumb("OTP verification failed", "warning", {
+            code: verifyError.code,
+            mode: "sign-up",
+          });
+          const { success } = handleOtpClerkError(verifyError);
+          if (!success) {
+            verifyingCodeRef.current = null;
+            setIsVerifying(false);
+            return;
+          }
+        }
+        if (signUp.status === "complete") {
+          authBreadcrumb("OTP verified", "info", { mode: "sign-up" });
+          setIsRedirecting(true);
+          await signUp.finalize({
+            navigate: ({ decorateUrl }) => {
+              window.location.href = decorateUrl(SUCCESS_REDIRECT);
+            },
+          });
+        } else {
+          verifyingCodeRef.current = null;
+          setOtpError(
+            "Sign-up could not be completed. Please try again or contact support."
+          );
+          setIsVerifying(false);
+        }
+      } catch {
+        verifyingCodeRef.current = null;
+        setOtpError("An unexpected error occurred. Please try again.");
+        setIsVerifying(false);
+      }
+    },
+    [signUp, handleOtpClerkError]
+  );
+
+  const onCodeChange = React.useCallback(
+    (value: string) => {
+      setOtpError(null);
+      if (value.length < 6) {
+        verifyingCodeRef.current = null;
+      }
+      setCode(value);
+      if (value.length !== 6) {
+        return;
+      }
+      if (verifyingCodeRef.current === value) {
+        return;
+      }
+      verifyingCodeRef.current = value;
+      void verifyOtp(value);
+    },
+    [verifyOtp]
+  );
+
+  const onResend = React.useCallback(async () => {
+    setIsResending(true);
+    setOtpError(null);
+    authBreadcrumb("OTP resend requested", "info", { mode: "sign-up" });
+    try {
+      const { error: sendError } = await authSpan(
+        "auth.otp.send",
+        { mode: "sign-up" },
+        () => signUp.verifications.sendEmailCode()
+      );
+      if (sendError) {
+        authBreadcrumb("OTP resend failed", "error", {
+          code: sendError.code,
+          mode: "sign-up",
+        });
+        handleOtpClerkError(sendError);
+      } else {
+        authBreadcrumb("OTP code resent", "info", { mode: "sign-up" });
+        toast.success("Verification code sent to your email");
+        setCode("");
+        verifyingCodeRef.current = null;
+      }
+    } catch {
+      setOtpError("An unexpected error occurred. Please try again.");
+    }
+    setIsResending(false);
+  }, [signUp, handleOtpClerkError]);
+
+  const onReset = React.useCallback(() => {
+    window.location.replace("/sign-up");
+  }, []);
+
+  const handleOAuth = React.useCallback(
+    async (strategy: OAuthStrategy) => {
+      if (oauthLoading) {
+        return;
+      }
+      if (!legalAccepted) {
+        setLegalError(
+          "You must accept the Terms of Service and Privacy Policy to continue."
+        );
+        return;
+      }
+      setLegalError(null);
+      setOauthLoading(true);
+      authBreadcrumb("OAuth sign-in initiated", "info", {
+        strategy,
+        mode: "sign-up",
+      });
+      try {
+        const { error: ssoError } = await authSpan(
+          "auth.oauth.initiate",
+          { mode: "sign-up", strategy },
+          () =>
+            signUp.sso({
+              strategy,
+              legalAccepted: true,
+              redirectCallbackUrl: "/sign-up/sso-callback",
+              redirectUrl: SUCCESS_REDIRECT,
+            })
+        );
+        if (ssoError) {
+          const mapped = mapOAuthClerkError(ssoError);
+          if (mapped.kind === "code" && mapped.errorCode === "waitlist") {
+            authBreadcrumb("OAuth blocked by waitlist (sso)", "warning", {
+              strategy,
+            });
+            handleWaitlist();
+            return;
+          }
+          if (mapped.kind === "redirect") {
+            window.location.href = mapped.target;
+            return;
+          }
+          if (mapped.kind === "inline") {
+            toast.error(mapped.message);
+            setOauthLoading(false);
+            return;
+          }
+        }
+        // On success, Clerk navigates to the IdP — control doesn't return.
+      } catch {
+        toast.error("An unexpected error occurred");
+        setOauthLoading(false);
+      }
+    },
+    [oauthLoading, legalAccepted, signUp, handleWaitlist]
+  );
 
   return (
     <div className="w-full max-w-md space-y-8">
-      {/* Header — only on email step */}
-      {step === "email" && !hasError && (
+      {view === "email" && !hasError && (
         <div className="text-center">
           <h1 className="font-medium font-pp text-3xl text-foreground">
-            {invitationTicket
-              ? "Accept Your Invitation"
-              : "Sign up for Lightfast"}
+            Sign up for Lightfast
           </h1>
         </div>
       )}
 
       <div className="space-y-4">
-        {/* Error display */}
         {hasError && (
           <ErrorBanner
-            backUrl={signUpBaseUrl}
+            backUrl="/sign-up"
             errorCode={errorCode}
-            message={error}
+            message={errorParam}
           />
         )}
 
-        {/* Step: email */}
-        {!hasError &&
-          step === "email" &&
-          (invitationTicket ? (
-            // Invitation flow — GitHub primary, email form secondary
-            <>
-              <OAuthButton
-                label="Continue with GitHub"
-                mode="sign-up"
-                strategy="oauth_github"
-                ticket={invitationTicket}
+        {!hasError && view === "email" && (
+          <>
+            <form className="space-y-4" onSubmit={onSubmit}>
+              <Input
+                autoComplete="email"
+                className="bg-background dark:bg-background"
+                disabled={submitting}
+                name="email"
+                onChange={(event) => setEmail(event.target.value)}
+                placeholder="Email Address"
+                required
+                size="lg"
+                type="email"
+                value={email}
               />
-              {env.NEXT_PUBLIC_VERCEL_ENV === "development" ? (
-                <OAuthButton
-                  label="Continue with Test IdP"
-                  mode="sign-up"
-                  strategy="oauth_custom_test_idp"
-                  ticket={invitationTicket}
+              <div className="flex items-start gap-3">
+                <Checkbox
+                  aria-describedby={legalError ? "legal-error" : undefined}
+                  checked={legalAccepted}
+                  className="mt-0.5"
+                  disabled={submitting}
+                  id="legal-accepted"
+                  onCheckedChange={(checked) => {
+                    setLegalAccepted(checked === true);
+                    if (checked === true) {
+                      setLegalError(null);
+                    }
+                  }}
                 />
-              ) : null}
-              <SeparatorWithText text="Or" />
-              <EmailForm action="sign-up" ticket={invitationTicket} />
-              {invitationExpiry && (
-                <p className="text-center text-muted-foreground text-xs">
-                  Invitation expires{" "}
-                  {invitationExpiry.toLocaleDateString("en-AU", {
-                    day: "numeric",
-                    month: "long",
-                    year: "numeric",
-                  })}
+                <label
+                  className="text-muted-foreground text-sm leading-relaxed"
+                  htmlFor="legal-accepted"
+                >
+                  I accept the{" "}
+                  <MicrofrontendLink
+                    className="text-foreground underline hover:text-foreground/80"
+                    href="/legal/terms"
+                    rel="noopener noreferrer"
+                    target="_blank"
+                  >
+                    Terms of Service
+                  </MicrofrontendLink>{" "}
+                  and{" "}
+                  <MicrofrontendLink
+                    className="text-foreground underline hover:text-foreground/80"
+                    href="/legal/privacy"
+                    rel="noopener noreferrer"
+                    target="_blank"
+                  >
+                    Privacy Policy
+                  </MicrofrontendLink>
+                </label>
+              </div>
+              {legalError && (
+                <p className="text-destructive text-sm" id="legal-error">
+                  {legalError}
                 </p>
               )}
-            </>
-          ) : (
-            // Standard waitlist sign-up — email form + GitHub secondary
-            <>
-              <EmailForm action="sign-up" ticket={null} />
+              <Button
+                className="w-full"
+                disabled={submitting}
+                size="lg"
+                type="submit"
+              >
+                {submitting ? (
+                  <Icons.spinner className="h-4 w-4 animate-spin" />
+                ) : (
+                  "Continue with Email"
+                )}
+              </Button>
+            </form>
+            <SeparatorWithText text="Or" />
+            <Button
+              className="w-full"
+              disabled={oauthLoading}
+              onClick={() => handleOAuth("oauth_github")}
+              size="lg"
+              variant="outline"
+            >
+              {oauthLoading ? (
+                <Icons.spinner className="mr-2 h-4 w-4 animate-spin" />
+              ) : (
+                <Icons.gitHub className="mr-2 h-4 w-4" />
+              )}
+              Continue with GitHub
+            </Button>
+            {env.NEXT_PUBLIC_VERCEL_ENV === "development" ? (
+              <Button
+                className="w-full"
+                disabled={oauthLoading}
+                onClick={() => handleOAuth("oauth_custom_test_idp")}
+                size="lg"
+                variant="outline"
+              >
+                {oauthLoading ? (
+                  <Icons.spinner className="mr-2 h-4 w-4 animate-spin" />
+                ) : (
+                  <Icons.gitHub className="mr-2 h-4 w-4" />
+                )}
+                Continue with Test IdP
+              </Button>
+            ) : null}
+            <div id="clerk-captcha" />
+          </>
+        )}
 
-              {/* Legal compliance */}
-              <p className="text-center text-muted-foreground text-sm">
-                By joining, you agree to our{" "}
-                <MicrofrontendLink
-                  className="text-foreground underline hover:text-foreground/80"
-                  href="/legal/terms"
-                  rel="noopener noreferrer"
-                  target="_blank"
-                >
-                  Terms of Service
-                </MicrofrontendLink>{" "}
-                and{" "}
-                <MicrofrontendLink
-                  className="text-foreground underline hover:text-foreground/80"
-                  href="/legal/privacy"
-                  rel="noopener noreferrer"
-                  target="_blank"
-                >
-                  Privacy Policy
-                </MicrofrontendLink>
-              </p>
-
-              <SeparatorWithText text="Or" />
-              <OAuthButton
-                label="Continue with GitHub"
-                mode="sign-up"
-                strategy="oauth_github"
-                ticket={null}
-              />
-              {env.NEXT_PUBLIC_VERCEL_ENV === "development" ? (
-                <OAuthButton
-                  label="Continue with Test IdP"
-                  mode="sign-up"
-                  strategy="oauth_custom_test_idp"
-                  ticket={null}
-                />
-              ) : null}
-            </>
-          ))}
-
-        {/* Step: code */}
-        {!hasError && step === "code" && email && (
-          <OTPIsland email={email} mode="sign-up" ticket={invitationTicket} />
+        {!hasError && view === "code" && (
+          <CodeVerificationUI
+            code={code}
+            email={submittedEmail}
+            inlineError={otpError}
+            isRedirecting={isRedirecting}
+            isResending={isResending}
+            isVerifying={isVerifying}
+            onCodeChange={onCodeChange}
+            onResend={onResend}
+            onReset={onReset}
+            title="Verify your email"
+          />
         )}
       </div>
 
-      {/* Sign In Link — only on email step */}
-      {step === "email" && !hasError && (
+      {view === "email" && !hasError && (
         <div className="text-center text-sm">
           <span className="text-muted-foreground">
             Already have an account?{" "}
